Move statistic to elastic-job-lite-tracing-rdb
